+++ xcube_resampling/__init__.py
@@ -19,6 +19,15 @@
 # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 # DEALINGS IN THE SOFTWARE.
 
-from .version import version
+from .version import __version__
+from .spatial import resample_in_space
+from .affine import affine_transform_dataset
+from .rectify import rectify_dataset
+from .reproject import reproject_dataset
 
-__version__ = version
+__all__ = [
+    "resample_in_space",
+    "affine_transform_dataset",
+    "rectify_dataset",
+    "reproject_dataset",
+]